Revert "move HttpClient metrics to public API"
This reverts commit 33376f74fc0ff90fc38b1d58879d5a58e855a83d.
