+++ ratpack-core/src/main/java/ratpack/http/client/internal/ChannelPoolStats.java
@@ -16,8 +16,6 @@
 
 package ratpack.http.client.internal;
 
-import ratpack.http.client.HostStats;
-
 public interface ChannelPoolStats {
 
   String getHost();
