+++ ratpack-core/src/main/java/ratpack/http/client/internal/HostStats.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 the original author or authors.
+ * Copyright 2016 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,8 @@
  * limitations under the License.
  */
 
-package ratpack.http.client;
+package ratpack.http.client.internal;
 
-/**
- * Http client pool metrics for a specific host.
- *
- * @since 1.9
- */
 public class HostStats {
   private final long activeConnectionCount;
   private final long idleConnectionCount;
