+++ ratpack-core/src/test/groovy/ratpack/http/client/internal/HttpClientStatsSpec.groovy
@@ -16,8 +16,6 @@
 
 package ratpack.http.client.internal
 
-import ratpack.http.client.HostStats
-import ratpack.http.client.HttpClientStats
 import ratpack.test.internal.BaseRatpackSpec
 
 class HttpClientStatsSpec extends BaseRatpackSpec {
