+++ ratpack-dropwizard-metrics/src/main/java/ratpack/dropwizard/metrics/internal/HttpClientMetrics.java
@@ -23,7 +23,8 @@ import com.google.inject.Singleton;
 import ratpack.dropwizard.metrics.DropwizardMetricsConfig;
 import ratpack.exec.ExecController;
 import ratpack.http.client.HttpClient;
-import ratpack.http.client.HttpClientStats;
+import ratpack.http.client.internal.DefaultHttpClient;
+import ratpack.http.client.internal.HttpClientStats;
 import ratpack.service.Service;
 import ratpack.service.StartEvent;
 
@@ -63,7 +64,7 @@ public class HttpClientMetrics implements Service, Runnable {
       boolean enabled = httpClientConfig.isEnabled();
       int pollingFrequency = httpClientConfig.getPollingFrequencyInSeconds();
 
-      if (enabled) {
+      if (enabled && httpClient instanceof DefaultHttpClient) {
         ExecController execController = event.getRegistry().get(ExecController.class);
         execController.getExecutor().scheduleAtFixedRate(this, 0, pollingFrequency, TimeUnit.SECONDS);
       }
@@ -72,7 +73,7 @@ public class HttpClientMetrics implements Service, Runnable {
 
   @Override
   public void run() {
-    HttpClientStats httpClientStats = httpClient.getHttpClientStats();
+    HttpClientStats httpClientStats = ((DefaultHttpClient) httpClient).getHttpClientStats();
     gauge(TOTAL_ACTIVE_CONNECTIONS).setValue(httpClientStats.getTotalActiveConnectionCount());
     gauge(TOTAL_IDLE_CONNECTIONS).setValue(httpClientStats.getTotalIdleConnectionCount());
     gauge(TOTAL_CONNECTIONS).setValue(httpClientStats.getTotalConnectionCount());
